Move BatchRequest import to fluent package
