+++ great_expectations/datasource/fluent/__init__.py
@@ -8,6 +8,7 @@ from great_expectations.datasource.fluent.interfaces import (
     Sorter,
     BatchMetadata,
 )
+from great_expectations.datasource.fluent.batch_request import BatchRequest
 from great_expectations.datasource.fluent.pandas_datasource import (
     PandasDatasource,
     _PandasDatasource,
