+++ great_expectations/datasource/fluent/data_asset/data_connector/file_path_data_connector.py
@@ -30,7 +30,7 @@ from great_expectations.datasource.fluent.data_asset.data_connector.regex_parser
 
 if TYPE_CHECKING:
     from great_expectations.core.batch import BatchDefinition
-    from great_expectations.datasource.fluent.batch_request import BatchRequest
+    from great_expectations.datasource.fluent import BatchRequest
 
 
 logger = logging.getLogger(__name__)
