+++ great_expectations/datasource/fluent/interfaces.py
@@ -685,7 +685,7 @@ class Batch(FluentBaseModel):
             BatchDefinition,
             BatchMarkers,
         )
-        from great_expectations.datasource.fluent.batch_request import BatchRequest
+        from great_expectations.datasource.fluent import BatchRequest
 
         super().update_forward_refs(
             BatchData=BatchData,
