+++ tests/datasource/fluent/data_asset/data_connector/test_filesystem_data_connector.py
@@ -7,7 +7,7 @@ import pytest
 
 from great_expectations.core import IDDict
 from great_expectations.core.batch import BatchDefinition
-from great_expectations.datasource.fluent.batch_request import BatchRequest
+from great_expectations.datasource.fluent import BatchRequest
 from great_expectations.datasource.fluent.data_asset.data_connector import (
     FilesystemDataConnector,
 )
