+++ tests/datasource/fluent/integration/integration_test_utils.py
@@ -10,7 +10,7 @@ from great_expectations.checkpoint import SimpleCheckpoint
 from great_expectations.checkpoint.types.checkpoint_result import CheckpointResult
 from great_expectations.data_context import AbstractDataContext
 from great_expectations.datasource.fluent import PandasDatasource
-from great_expectations.datasource.fluent.batch_request import BatchRequest
+from great_expectations.datasource.fluent import BatchRequest
 from great_expectations.datasource.fluent.interfaces import (
     DataAsset,
     Datasource,
